[project @ robey@lag.net-20070210235828-2a24iw2xqe4lnf42] patch from mpool to fix a python 2.5 warning: stat() returns floats for times and we want to encode them as ints.

+++ paramiko/sftp_attr.py
@@ -136,8 +136,9 @@ class SFTPAttributes (object):
         if self._flags & self.FLAG_PERMISSIONS:
             msg.add_int(self.st_mode)
         if self._flags & self.FLAG_AMTIME:
-            msg.add_int(self.st_atime)
-            msg.add_int(self.st_mtime)
+            # throw away any fractional seconds
+            msg.add_int(long(self.st_atime))
+            msg.add_int(long(self.st_mtime))
         if self._flags & self.FLAG_EXTENDED:
             msg.add_int(len(self.attr))
             for key, val in self.attr.iteritems():
